finalize move to external lib
