+++ package.json
@@ -22,7 +22,7 @@
         "test:debug": "npx --node-arg=--inspect jest --runInBand"
     },
     "engines": {
-        "node": ">=12.13.0",
+        "node": ">=10.4.0",
         "npm": ">=5.6.0"
     },
     "repository": {
@@ -39,9 +39,7 @@
         "@org-formation/tombok": "^0.0.1",
         "autobind-decorator": "^2.4.0",
         "class-transformer": "^0.3.1",
-        "piscina": "^1.6.3",
         "reflect-metadata": "^0.1.13",
-        "serialize-error": "^7.0.1",
         "string.prototype.replaceall": "^1.0.3",
         "uuid": "^7.0.2"
     },
@@ -52,6 +50,7 @@
         "@types/uuid": "^7.0.0",
         "@typescript-eslint/eslint-plugin": "^4.8.1",
         "@typescript-eslint/parser": "^4.8.1",
+        "aws-sdk": "~2.712.0",
         "eslint": "^7.13.0",
         "eslint-config-prettier": "^6.15.0",
         "eslint-import-resolver-node": "^0.3.3",
@@ -61,10 +60,11 @@
         "eslint-plugin-prettier": "^3.1.3",
         "jest": "^26.6.3",
         "jest-circus": "^26.6.3",
-        "prettier": "^2.1.2",
+        "prettier": "^2.2.0",
         "ts-jest": "^26.4.4",
         "ts-node": "^9.0.0",
-        "typescript": "^4.0.5"
+        "typescript": "^4.1.2",
+        "worker-pool-aws-sdk": "^0.0.2"
     },
     "optionalDependencies": {
         "aws-sdk": "~2.712.0"
